--- javav2/example_code/dynamodb/src/main/java/com/example/dynamodb/ScenarioPartiQ.java
@@ -1,345 +1,0 @@
-// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-package com.example.dynamodb;
-
-// snippet-start:[dynamodb.java2.scenario.partiql.import]
-import com.fasterxml.jackson.databind.JsonNode;
-import software.amazon.awssdk.core.waiters.WaiterResponse;
-import software.amazon.awssdk.regions.Region;
-import software.amazon.awssdk.services.dynamodb.DynamoDbClient;
-import software.amazon.awssdk.services.dynamodb.model.KeySchemaElement;
-import software.amazon.awssdk.services.dynamodb.model.DynamoDbException;
-import software.amazon.awssdk.services.dynamodb.model.CreateTableRequest;
-import software.amazon.awssdk.services.dynamodb.model.ProvisionedThroughput;
-import software.amazon.awssdk.services.dynamodb.model.DescribeTableRequest;
-import software.amazon.awssdk.services.dynamodb.model.DescribeTableResponse;
-import software.amazon.awssdk.services.dynamodb.model.AttributeValue;
-import software.amazon.awssdk.services.dynamodb.model.CreateTableResponse;
-import software.amazon.awssdk.services.dynamodb.model.KeyType;
-import software.amazon.awssdk.services.dynamodb.model.ExecuteStatementResponse;
-import software.amazon.awssdk.services.dynamodb.model.AttributeDefinition;
-import software.amazon.awssdk.services.dynamodb.model.DeleteTableRequest;
-import software.amazon.awssdk.services.dynamodb.model.ExecuteStatementRequest;
-import software.amazon.awssdk.services.dynamodb.waiters.DynamoDbWaiter;
-import com.fasterxml.jackson.core.JsonFactory;
-import com.fasterxml.jackson.core.JsonParser;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.node.ObjectNode;
-import java.io.File;
-import java.io.IOException;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Iterator;
-// snippet-end:[dynamodb.java2.scenario.partiql.import]
-
-/**
- * Before running this Java V2 code example, set up your development
- * environment, including your credentials.
- *
- * For more information, see the following documentation topic:
- *
- * https://docs.aws.amazon.com/sdk-for-java/latest/developer-guide/get-started.html
- *
- * This Java example performs the following tasks:
- *
- * 1. Creates the Amazon DynamoDB movie table with a partition and sort key.
- * 2. Puts data into the Amazon DynamoDB table from a JSON document.
- * 3. Adds a new item.
- * 4. Gets an item by the composite key (the Partition key and Sort key).
- * 5. Updates an item.
- * 6. Uses a Scan to query items using the Enhanced client.
- * 7. Queries all items where the year is 2013 using the Enhanced Client.
- * 8. Deletes the table.
- *
- * To see another code example with more options using PartiQL and Batch
- * commands, see the ScenarioPartiQBatch code example.
- */
-
-// snippet-start:[dynamodb.java2.scenario.partiql.main]
-public class ScenarioPartiQ {
-    public static void main(String[] args) throws IOException {
-        final String usage = """
-
-                Usage:
-                    <fileName>
-
-                Where:
-                    fileName - The path to the moviedata.json file that you can download from the Amazon DynamoDB Developer Guide.
-                """;
-
-        if (args.length != 1) {
-            System.out.println(usage);
-            System.exit(1);
-        }
-
-        String fileName = args[0];
-        String tableName = "MoviesPartiQ";
-        Region region = Region.US_EAST_1;
-        DynamoDbClient ddb = DynamoDbClient.builder()
-                .region(region)
-                .build();
-
-        System.out.println(
-                "******* Creating an Amazon DynamoDB table named MoviesPartiQ with a key named year and a sort key named title.");
-        createTable(ddb, tableName);
-
-        System.out.println("******* Loading data into the MoviesPartiQ table.");
-        loadData(ddb, fileName);
-
-        System.out.println("******* Getting data from the MoviesPartiQ table.");
-        getItem(ddb);
-
-        System.out.println("******* Putting a record into the MoviesPartiQ table.");
-        putRecord(ddb);
-
-        System.out.println("******* Updating a record.");
-        updateTableItem(ddb);
-
-        System.out.println("******* Querying the movies released in 2013.");
-        queryTable(ddb);
-
-        System.out.println("******* Deleting the Amazon DynamoDB table.");
-        deleteDynamoDBTable(ddb, tableName);
-        ddb.close();
-    }
-
-    public static void createTable(DynamoDbClient ddb, String tableName) {
-        DynamoDbWaiter dbWaiter = ddb.waiter();
-        ArrayList<AttributeDefinition> attributeDefinitions = new ArrayList<>();
-
-        // Define attributes.
-        attributeDefinitions.add(AttributeDefinition.builder()
-                .attributeName("year")
-                .attributeType("N")
-                .build());
-
-        attributeDefinitions.add(AttributeDefinition.builder()
-                .attributeName("title")
-                .attributeType("S")
-                .build());
-
-        ArrayList<KeySchemaElement> tableKey = new ArrayList<>();
-        KeySchemaElement key = KeySchemaElement.builder()
-                .attributeName("year")
-                .keyType(KeyType.HASH)
-                .build();
-
-        KeySchemaElement key2 = KeySchemaElement.builder()
-                .attributeName("title")
-                .keyType(KeyType.RANGE) // Sort
-                .build();
-
-        // Add KeySchemaElement objects to the list.
-        tableKey.add(key);
-        tableKey.add(key2);
-
-        CreateTableRequest request = CreateTableRequest.builder()
-                .keySchema(tableKey)
-                .provisionedThroughput(ProvisionedThroughput.builder()
-                        .readCapacityUnits(new Long(10))
-                        .writeCapacityUnits(new Long(10))
-                        .build())
-                .attributeDefinitions(attributeDefinitions)
-                .tableName(tableName)
-                .build();
-
-        try {
-            CreateTableResponse response = ddb.createTable(request);
-            DescribeTableRequest tableRequest = DescribeTableRequest.builder()
-                    .tableName(tableName)
-                    .build();
-
-            // Wait until the Amazon DynamoDB table is created.
-            WaiterResponse<DescribeTableResponse> waiterResponse = dbWaiter.waitUntilTableExists(tableRequest);
-            waiterResponse.matched().response().ifPresent(System.out::println);
-            String newTable = response.tableDescription().tableName();
-            System.out.println("The " + newTable + " was successfully created.");
-
-        } catch (DynamoDbException e) {
-            System.err.println(e.getMessage());
-            System.exit(1);
-        }
-    }
-
-    // Load data into the table.
-    public static void loadData(DynamoDbClient ddb, String fileName) throws IOException {
-
-        String sqlStatement = "INSERT INTO MoviesPartiQ VALUE {'year':?, 'title' : ?, 'info' : ?}";
-        JsonParser parser = new JsonFactory().createParser(new File(fileName));
-        com.fasterxml.jackson.databind.JsonNode rootNode = new ObjectMapper().readTree(parser);
-        Iterator<JsonNode> iter = rootNode.iterator();
-        ObjectNode currentNode;
-        int t = 0;
-        List<AttributeValue> parameters = new ArrayList<>();
-        while (iter.hasNext()) {
-
-            // Add 200 movies to the table.
-            if (t == 200)
-                break;
-            currentNode = (ObjectNode) iter.next();
-
-            int year = currentNode.path("year").asInt();
-            String title = currentNode.path("title").asText();
-            String info = currentNode.path("info").toString();
-
-            AttributeValue att1 = AttributeValue.builder()
-                    .n(String.valueOf(year))
-                    .build();
-
-            AttributeValue att2 = AttributeValue.builder()
-                    .s(title)
-                    .build();
-
-            AttributeValue att3 = AttributeValue.builder()
-                    .s(info)
-                    .build();
-
-            parameters.add(att1);
-            parameters.add(att2);
-            parameters.add(att3);
-
-            // Insert the movie into the Amazon DynamoDB table.
-            executeStatementRequest(ddb, sqlStatement, parameters);
-            System.out.println("Added Movie " + title);
-
-            parameters.remove(att1);
-            parameters.remove(att2);
-            parameters.remove(att3);
-            t++;
-        }
-    }
-
-    public static void getItem(DynamoDbClient ddb) {
-
-        String sqlStatement = "SELECT * FROM MoviesPartiQ where year=? and title=?";
-        List<AttributeValue> parameters = new ArrayList<>();
-        AttributeValue att1 = AttributeValue.builder()
-                .n("2012")
-                .build();
-
-        AttributeValue att2 = AttributeValue.builder()
-                .s("The Perks of Being a Wallflower")
-                .build();
-
-        parameters.add(att1);
-        parameters.add(att2);
-
-        try {
-            ExecuteStatementResponse response = executeStatementRequest(ddb, sqlStatement, parameters);
-            System.out.println("ExecuteStatement successful: " + response.toString());
-
-        } catch (DynamoDbException e) {
-            System.err.println(e.getMessage());
-            System.exit(1);
-        }
-    }
-
-    public static void putRecord(DynamoDbClient ddb) {
-
-        String sqlStatement = "INSERT INTO MoviesPartiQ VALUE {'year':?, 'title' : ?, 'info' : ?}";
-        try {
-            List<AttributeValue> parameters = new ArrayList<>();
-
-            AttributeValue att1 = AttributeValue.builder()
-                    .n(String.valueOf("2020"))
-                    .build();
-
-            AttributeValue att2 = AttributeValue.builder()
-                    .s("My Movie")
-                    .build();
-
-            AttributeValue att3 = AttributeValue.builder()
-                    .s("No Information")
-                    .build();
-
-            parameters.add(att1);
-            parameters.add(att2);
-            parameters.add(att3);
-
-            executeStatementRequest(ddb, sqlStatement, parameters);
-            System.out.println("Added new movie.");
-
-        } catch (DynamoDbException e) {
-            System.err.println(e.getMessage());
-            System.exit(1);
-        }
-    }
-
-    public static void updateTableItem(DynamoDbClient ddb) {
-
-        String sqlStatement = "UPDATE MoviesPartiQ SET info = 'directors\":[\"Merian C. Cooper\",\"Ernest B. Schoedsack' where year=? and title=?";
-        List<AttributeValue> parameters = new ArrayList<>();
-        AttributeValue att1 = AttributeValue.builder()
-                .n(String.valueOf("2013"))
-                .build();
-
-        AttributeValue att2 = AttributeValue.builder()
-                .s("The East")
-                .build();
-
-        parameters.add(att1);
-        parameters.add(att2);
-
-        try {
-            executeStatementRequest(ddb, sqlStatement, parameters);
-
-        } catch (DynamoDbException e) {
-            System.err.println(e.getMessage());
-            System.exit(1);
-        }
-        System.out.println("Item was updated!");
-    }
-
-    // Query the table where the year is 2013.
-    public static void queryTable(DynamoDbClient ddb) {
-        String sqlStatement = "SELECT * FROM MoviesPartiQ where year = ? ORDER BY year";
-        try {
-
-            List<AttributeValue> parameters = new ArrayList<>();
-            AttributeValue att1 = AttributeValue.builder()
-                    .n(String.valueOf("2013"))
-                    .build();
-            parameters.add(att1);
-
-            // Get items in the table and write out the ID value.
-            ExecuteStatementResponse response = executeStatementRequest(ddb, sqlStatement, parameters);
-            System.out.println("ExecuteStatement successful: " + response.toString());
-
-        } catch (DynamoDbException e) {
-            System.err.println(e.getMessage());
-            System.exit(1);
-        }
-    }
-
-    public static void deleteDynamoDBTable(DynamoDbClient ddb, String tableName) {
-
-        DeleteTableRequest request = DeleteTableRequest.builder()
-                .tableName(tableName)
-                .build();
-
-        try {
-            ddb.deleteTable(request);
-
-        } catch (DynamoDbException e) {
-            System.err.println(e.getMessage());
-            System.exit(1);
-        }
-        System.out.println(tableName + " was successfully deleted!");
-    }
-
-    private static ExecuteStatementResponse executeStatementRequest(DynamoDbClient ddb, String statement,
-            List<AttributeValue> parameters) {
-        ExecuteStatementRequest request = ExecuteStatementRequest.builder()
-                .statement(statement)
-                .parameters(parameters)
-                .build();
-
-        return ddb.executeStatement(request);
-    }
-
-    private static void processResults(ExecuteStatementResponse executeStatementResult) {
-        System.out.println("ExecuteStatement successful: " + executeStatementResult.toString());
-    }
-}
-// snippet-end:[dynamodb.java2.scenario.partiql.main]
