+++ javav2/example_code/dynamodb/src/test/java/DynamoDBTest.java
@@ -10,8 +10,8 @@ import com.example.dynamodb.GetItem;
 import com.example.dynamodb.ListTables;
 import com.example.dynamodb.PutItem;
 import com.example.dynamodb.Query;
-import com.example.dynamodb.Scenario;
-import com.example.dynamodb.ScenarioPartiQ;
+import com.example.dynamodb.scenario.Scenario;
+import com.example.dynamodb.scenario.ScenarioPartiQ;
 import com.example.dynamodb.SyncPagination;
 import com.example.dynamodb.UpdateItem;
 import com.example.dynamodb.UpdateTable;
